+++ ref/gl/gl_local.h
@@ -227,9 +227,9 @@ typedef struct
 	cl_entity_t *viewent;
 	dlight_t *elights;
 	byte *texgammatable;
-	uint *lightgammatable;
-	uint *lineargammatable;
-	uint *screengammatable;
+	uint16_t *lightgammatable;
+	uint16_t *lineargammatable;
+	uint16_t *screengammatable;
 
 	uint max_entities;
